Remove high level helpers and place them to clients
