+++ sdk/storage/storage-blob/src/index.ts
@@ -27,9 +27,7 @@ export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
 export * from "./credentials/SharedKeyCredential";
 export * from "./credentials/TokenCredential";
-export * from "./highlevel.browser";
 export * from "./highlevel.common";
-export * from "./highlevel.node";
 export { IPRange } from "./IPRange";
 export { Range } from "./Range";
 export * from "./Pipeline";
