+++ sdk/storage/storage-blob/test/browser/highlevel.browser.spec.ts
@@ -1,7 +1,6 @@
 import * as assert from "assert";
 
 import { Aborter } from "../../src/Aborter";
-import { uploadBrowserDataToBlockBlob } from "../../src/highlevel.browser";
 import {
   arrayBufferEqual,
   blobToArrayBuffer,
@@ -50,7 +49,7 @@ describe("Highelvel", () => {
     const aborter = Aborter.timeout(1);
 
     try {
-      await uploadBrowserDataToBlockBlob(tempFile1, blockBlobClient, {
+      await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile1, {
         abortSignal: aborter
       });
       assert.fail();
@@ -63,7 +62,7 @@ describe("Highelvel", () => {
     const aborter = Aborter.timeout(1);
 
     try {
-      await uploadBrowserDataToBlockBlob(tempFile2, blockBlobClient, {
+      await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile2, {
         abortSignal: aborter,
         blockSize: 4 * 1024 * 1024,
         parallelism: 2
@@ -79,7 +78,7 @@ describe("Highelvel", () => {
     const aborter = Aborter.none;
 
     try {
-      await uploadBrowserDataToBlockBlob(tempFile1, blockBlobClient, {
+      await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile1, {
         abortSignal: aborter,
         blockSize: 4 * 1024 * 1024,
         parallelism: 2,
@@ -98,7 +97,7 @@ describe("Highelvel", () => {
     const aborter = Aborter.none;
 
     try {
-      await uploadBrowserDataToBlockBlob(tempFile2, blockBlobClient, {
+      await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile2, {
         abortSignal: aborter,
         blockSize: 4 * 1024 * 1024,
         parallelism: 2,
@@ -113,7 +112,7 @@ describe("Highelvel", () => {
   });
 
   it("uploadBrowserDataToBlockBlob should success when blob < BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES", async () => {
-    await uploadBrowserDataToBlockBlob(tempFile2, blockBlobClient, {
+    await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile2, {
       blockSize: 4 * 1024 * 1024,
       parallelism: 2
     });
@@ -126,7 +125,7 @@ describe("Highelvel", () => {
   });
 
   it("uploadBrowserDataToBlockBlob should success when blob < BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES and configured maxSingleShotSize", async () => {
-    await uploadBrowserDataToBlockBlob(tempFile2, blockBlobClient, {
+    await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile2, {
       blockSize: 512 * 1024,
       maxSingleShotSize: 0
     });
@@ -147,7 +146,7 @@ describe("Highelvel", () => {
       return;
     }
 
-    await uploadBrowserDataToBlockBlob(tempFile1, blockBlobClient, {
+    await blockBlobClient.uploadBrowserDataToBlockBlob(tempFile1, {
       blockSize: 4 * 1024 * 1024,
       parallelism: 2
     });
